Fix casts

+++ tsMuxer/bufferedFileReader.h
@@ -2,7 +2,6 @@
 #define BUFFERED_FILE_READER_H_
 
 #include <fs/file.h>
-#include <fs/systemlog.h>
 #include <types/types.h>
 
 #include "BufferedReader.h"
@@ -16,15 +15,9 @@ class FileListIterator final : public FileNameIterator
     FileListIterator() : m_index(0) {}
     ~FileListIterator() override = default;
 
-    std::string getNextName() override
-    {
-        return ++m_index < m_files.size() ? m_files[m_index] : "";
-    }
+    std::string getNextName() override { return ++m_index < m_files.size() ? m_files[m_index] : ""; }
 
-    void addFile(const std::string& fileName)
-    {
-        m_files.push_back(fileName);
-    }
+    void addFile(const std::string& fileName) { m_files.push_back(fileName); }
 
    private:
     std::vector<std::string> m_files;
@@ -39,14 +32,11 @@ struct FileReaderData final : ReaderData
 
     ~FileReaderData() override = default;
 
-    uint32_t readBlock(uint8_t* buffer, const int max_size) override { return m_file.read(buffer, max_size); }
+    uint32_t readBlock(uint8_t* buffer, const uint32_t max_size) override { return m_file.read(buffer, max_size); }
 
     bool openStream() override;
     bool closeStream() override { return m_file.close(); }
-    bool incSeek(const int64_t offset) override
-    {
-        return m_file.seek(offset, File::SeekMethod::smCurrent) != static_cast<uint64_t>(-1);
-    }
+    bool incSeek(const int64_t offset) override { return m_file.seek(offset, File::SeekMethod::smCurrent) != -1; }
 
     File m_file;
     uint32_t m_fileHeaderSize;
@@ -57,9 +47,8 @@ class BufferedFileReader final : public BufferedReader
    public:
     BufferedFileReader(uint32_t blockSize, uint32_t allocSize = 0, uint32_t prereadThreshold = 0);
 
-    bool openStream(uint32_t readerID, const char* streamName, int pid = 0,
-                    const CodecInfo* codecInfo = nullptr) override;
-    bool gotoByte(uint32_t readerID, uint64_t seekDist) override;
+    bool openStream(int readerID, const char* streamName, int pid = 0, const CodecInfo* codecInfo = nullptr) override;
+    bool gotoByte(int readerID, int64_t seekDist) override;
 
    protected:
     ReaderData* intCreateReader() override { return new FileReaderData(m_blockSize, m_allocSize); }
